apartment/requirements.txt
boto3==1.2.4
botocore==1.3.28
Django==1.9.2
docutils==0.12
jmespath==0.9.0
python-dateutil==2.4.2
six==1.10.0
wheel==0.29.0
djangorestframework==3.3.2
markdown==2.6.5
